// src/components/Dumb/Text/text.types.ts
export type CTextAs = "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "p";

export type CTextSize = "4xl" | "3xl" | "2xl" | "xl" | "lg" | "md" | "sm" | "xs";

export type CTextColors =
  | "primary"
  | "secondary"
  | "white"
  | "dark"
  | "dark2"
  | "medium"
  | "light"
  | "danger";

export type CFontWeights = "black" | "semibold" | "bold" | "light" | "regular" | "medium";

export type CFontFamily = "Asap Condensed" | "Helvetica";

export type CTextPosition = "text-center" | "text-left" | "text-right";